backend/requirements.txt
asgiref==3.2.10
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
constantly==15.1.0
cryptography==3.1.1
dateparser==0.7.6
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
Jinja2==2.11.2
MarkupSafe==1.1.1
psutil==5.7.2
pushbullet.py==0.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
python-magic==0.4.18
pytz==2020.1
regex==2020.10.15
requests==2.24.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
urllib3==1.25.10
websocket-client==0.57.0
zope.interface==5.1.2
pycoingecko
cryptowatch-sdk
ccxt
coinapi.rest.v1
